Return a list of un-run migrations if the migration state table is absent
With this patch calling `diff` will create the database (should it
really?) and check what migrations are run.

If the database doesn't exist, or the table doesn't exist (we don't
create that) then we answer with a new iterator with tuples of (false,
TheMigration) to signal to the caller which migrations are run or not.

Because of the existing tests we paired on at the time, we had an "it
implicitly creates the DB" test, that felt reasonable at the time, and
maybe I can be convinced that it still does. The challenge I faced, is
that to create the migration state *table* I don't want to hard-code
the create table statement, I want to run the first migrations, however
I don't want to do writing changes when just running a diff.

Having gotten this far, what I think I want to do is move next to "up()",
which is going to use diff() to find out what has to run, and I am going
to move CREATE DATABASE and CREATE TABLE *both* to a bootstrap
migration, and solve the templating language challenges there. That way
diff() becomes a pure read-only function, and if either the schema or
the table, or both don't exist, at the first sign of missing stuff, we
return this new iterator with all values in the state tuple set to
false.

Extra notes after this commit:

- true/false on the MigrationStateTuple seems slightly ambiguous, is it
"true, is run", or "false, hasn't run yet", or "false, don't need to
run. Think an enum of Pending vs. Applied would be better here (or
something like this)

- the lifetime parameters! I was able to make something work, and I even
think it might be correct. I think in the end this reads "the input, and
the Ok() return value must have the same lifetime"