Documentation for config types, and reserved words (`cargo doc --open` is nice)
This splits the `crate::runner::Runner` into `Runner` and `StateStore`
(note: both live in the same file for now, probably they shouldn't)

Ths reason for the split is that StateStore needs a whole Config (with
all the runner configs) so when they list, diff, and up migrations, they
can use the full complement of configs to instantiate a runner for that
isolated case.

The MariaDB struct could probably be split too, there are now four
fields on it, 3 used in one 2 used in the other. Probably it makes sense
to split it.

Right now Runner<Error =T> is dynamic on the trait, which means that
each of MySQL, PostgreSQL and other stateful stores would have to
implement the look-up and creation map. It may make sense to have
`Runner` return static errors (not MariaDB, PostgreSQL specific ones,
etc) in order that one generic implementation could be used.

There is confusion in this branch now between MigrationStep.runner
(which runner was identified in the parsing of the migration) and .. the
concept of a _configured_ runner. Probably this is just oversight on my
part, but the MigrationStep knows it wants to run with MariaDb, but it
does not know which _configuration_ (username, password, port, etc) it
wants to use. Migration also doesn't know. Probably this means that we
need to look at the migration parsing again real quick and figure out
how to link them up.

I'm leaning towards something like `...myconfigentry.sql` being
interpreted as "I expect a config called "myconfigentry" which has a
runner which has ".sql" in the list of extensions. And then this
"myconfigentry" (as a standalone str[ing]) is remembered along side the
migration. This would make it easier to find a runner. Right now all
runners are MariaDB all the time, hard-coded. Taclking this also forces
me to finally confront the ambiguity in the file naming scheme.

Databases are not properly cleaned up, I need mitre and mariadb configs
on the test configuration now, and the Drop trait implementation is only
dropping one database, we need to do a severe round of refactoring
before this can merge.

Other small changes ctor for ensuring that env_logger::init is called. I
made the mistake of looking into how ctor work, and it is cursed: https://github.com/mmastrac/rust-ctor#under-the-hood