Partially working MySQL MigrationStorage including bootstrapping and reset
The bootstrapping and reset have highlighted a lot of potential overlap
with the Migration runner (e.g parsing/loading). I think the concept of
built-in migrations, as neat as it seemed is not suitable, although the
pre-processing of a Migration given a Config/Context (mustache, parsing
variables and interpolation) seems valuable as a standalone thing.

Lots of the bootstrapping (namely interpolating the database name, which
is dynamic) overlaps a lot with a template context and other things.
Probably the answer is that a config has a "to template context" method
that, or there's a HasTemplateContext trait that stacks somehow (mysql
driver context for the two table names, plus config context which gives
access to the configuration for dbname and ports/etc). To be revisited.

The "built-in" migrations idea also seems less important since I
discovered that probably it's weird to see migrations you didn't write
show up in the `ls`, and if now because of the new MigrationStorage
trait, it becomes easier to write arbitrary storages which are not
runners, which I think might be true, then it doesn't make sense to bind
set-up of the MigrationStorage to some concept of a runnable migration.

Maybe we keep the concept of build-in migration (or, let's say, we don't
rule it out) I could imagine improving it with some MultiMigrationList
which concatenates multiple MigrationList. But so far no pressure to
decide anything. Built-in migration code is still here, still tested,
and still works.

I briefly experimented with having the MigrationStorage trait take a
&Migration, as I realized that iterating over Migration.Steps
(HashMap<Direction,MigrationStep>) requires a borrow because I presume
it is stateful. Anyway, simply cloning the steps before iterating was
less painful than implementing Copy for all the Migration object graph.

It also raised a question a bit, the InMemoryMigration storage
implementation, should this return the *same memory object* (now,
it will, I think because it takes Migration, not &Migration), which will
likely cause ownership/borrow problems.

We can revisit this decision any time too. It's infeasible (I think)
right now to implement the trait in terms of &Migration because we
cannot derive, and I do not want to write a Copy implementation for the
Migration, MigrationSteps<Direction, MigrationStep>, MigrationStep, and
Direction types. It can be derived for most, but not the HashMap since
you can't use derive on type aliases. I'm sure there's an "ok" solution,
but for now I'm skirting the issue a bit.