Make more-or-less proper use of Iterator trait.
For most of the Engine we simply do not care what _type_ of thing we
have (vec, tree, some custom implementation, etc) we simply want to
capture the notion of _iterable_.

For this we can use the `Iterator<Item = T>` trait. The trait is not a
real type, so eventually we have to return something that is
Iterator<T>. However in our own traits (MigrationStorage, MigrationList)
we can say "migration list answers with Iterator<Item=Mirgation>" and
rely on different implemetations to do different things.

In-Memory migrations return std::vec::IntoIter<Migration> because
IntoIter is a *type* not a trait.

Other implementations might return something like Map<Migration> if we
have a map over some other collection (e.g the filesystem one
filter_map'ping paths into Some(Migration). The type of that is probably
something like Map<ignore::walk::DirEntry<T>> or similar.

For the MySQL implementation there is probably a way to do an
Iterator<T> without needing to collect all the results up-front using
some kind of row cursor over the results table. If anyone ever has
enough migrations to warrant optimizing/rewriting that, they can pay me
to rewrite it. For now it returns a stc::vec::IntoIter<Migration> like
all the other implementations.

Three reasons to keep the MigrationList trait which is _not_ an
Iterator<T> but returns one in the all() function:

- First all() returns Result<Iterator<Migration>, E> so we can know if
  we failed, Iterator<T> doesn't have a great way to express that with
  just an Iterator<T>, we'd just get an empty collection.

- Secondly ownership can/would be weird. If we wanted to share a mysql
  conn (_if_ indeed we do) then the iterator needing an indefinite (we
  don't know when it would be drained) mutable borrow on the
  `impl MigrationList for MySQL` self would cause a problem because the
  `self` would also probably want to do something with that `&mut self`
  (specifically the connection on it, which requries a mutable borrow to
  do anything on the connection).

- By having `all()` return a new iterator, in theory this could use a
  2nd mysql connection, and _generally_ keeping `all()` as trait method,
  makes it easier to write a type which is both `MigrationList` and
  `MigrationStorage`. If the MySQL list+store was itself both
  Iterator<Migration> and MigrationStorage then it becomes really fuzzy
  what happens semantically if you take some values with next() and then
  push another one, does this restart the iterator, reset the position,
  re-run the query, etc, etc? So for that reason, all() gives us back
  everything known at the time that functuon is called.

In the end, I think this basically wraps up the Iterator<T>,
IntoIterator<T> rabbit hole, and I think by now I almost understand the
type/trait system, and how trait bounds and generics work (at least as
far as here) but I'm not thrilled with the ergnomics of the thing.I
_hoped_ that the `impl MigrationList for X { type Iterator = ....` would
be inferred by the compiler, and, indeed there are unstable features
(which don't look like they arrive any time soon) coming which _might_
make this better, but needing to hard-code this _somehow_ makes me feel
like this undermined the usefulness of this generics trait binding.