Second pass implementing advice from @chris-morgan thanks to his kind comments in https://github.com/leehambley/mitre/pull/32#commitcomment-51554080
Using Box<(dyn Iterator<Item=Migration)> works much more cleanly in the
cases I was concerned about (migration_storage::from_config, and the
test fixtures).

The associated types turn out not to be important, as Chris Morgan
mentioned in his notes, they're interesting if a trait would have
different iteraable types, and _maybe_ it's worth having that in some
cases, but given that this library does I/O and database transformations
I feel confident saying that the overhead of a Box<T> is well within
tolerance.

Somehow the end-state of the code isn't really reflecting all the pain
we went through to get it here, but I'm actually really satisfied with
how this all panned out.

The lifetime annotations on the from_disk() implementation are still a
bit spotty, both 'a and 'b are there for reasons I only _partly_
understand, and I think 'a can go away if refactored gently, which means
I only have to wrap my brain around 'b and work from there.