WIP on Engine
I seem to be confused between Driver.apply() and Runner.apply() the
former expects a &Migration to apply the correct steps (up or change)
where the latter seems to expect already an individual &MigrationStep.

I believe that `Runner` was on the way out with `Driver` replacing it,
but I don't entirely remember.

The API of Driver seems more compatible with the new
MigrationList+MigrationStorage traits, but I can hardly be sure.

I will pick up here tomorrow, and try and make this cleaner.